sales final/sales/sales/sales/requirements.txt
flask>=3.0.0
tinydb>=4.8.0
python-dateutil>=2.8.2
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
openpyxl>=3.1.0
